refactor: compListSection 생성 compList 분리
