+++ components/module/CompListSection/CompList/CompCard/CompStatusInfo/CompStatusInfo.tsx
